Feat: #2 로그인, 회원가입 기능 구현
- JWT
- 로그인, 회원가입, 로그아웃 기능 구현
- Vuex 설정
