+++ client/src/views/SignIn.vue
@@ -1,122 +1,193 @@
 <template>
-    <section>
-        <div class="login-box">
-            <h2>Login</h2>
-            <form>
-                <div class="user-box">
-                    <input type="text" name="" required="">
-                    <label>Username</label>
-                </div>
-                <div class="user-box">
-                    <input type="password" name="" required="">
-                    <label>Password</label>
-                </div>
-                <a href="#">
-                    <span></span>
-                    <span></span>
-                    <span></span>
-                    <span></span>
-                    Submit
-                </a>
-            </form>
+  <section>
+    <div class="login-box">
+      <div v-if="!isSignUp">
+        <h2>Sign in</h2>
+        <div class="user-box">
+          <input type="text" v-model.trim="form.email" required=true  maxlength="100">
+          <label>email</label>
         </div>
-    </section>
+        <div class="user-box">
+          <input type="password" v-model="form.password" required=true  maxlength="100">
+          <label>password</label>
+        </div>
+        <button class="login-btn" @click="signin()">signin</button>
+        <button class="login-btn" @click="changeForm()">signup</button>
+      </div>
+      <div v-if="isSignUp">
+        <h2>Sign up</h2>
+        <div class="user-box">
+          <input type="text" v-model.trim="form.email" required=true maxlength="100">
+          <label>email</label>
+        </div>
+        <div class="user-box">
+          <input type="text" v-model.trim="form.name" required=true maxlength="50">
+          <label>userName</label>
+        </div>
+        <div class="user-box">
+          <input type="password" v-model="form.password" required=true maxlength="100">
+          <label>password</label>
+        </div>
+        <div class="user-box">
+          <input type="password" v-model="password2" required=true maxlength="100">
+          <label>confirm password</label>
+        </div>
+        <button class="login-btn" @click="changeForm()">signin</button>
+        <button class="login-btn" @click="signup()">signup</button>
+      </div>
+    </div>
+  </section>
 </template>
 <script>
+import axios from 'axios';
+
+export default {
+  data() {
+    return {
+      isSignUp: false,
+      password2: null,
+      form: {
+        email: null,
+        password: null,
+        name: null,
+      }
+    }
+  },
+  methods: {
+    changeForm() {
+      this.isSignUp = !this.isSignUp;
+      this.form.password = null;
+      this.password2 = null;
+    },
+    async signin() {
+      if (this.form.email == null || this.form.password == null) {
+          alert("입력하지 않은 칸이 있습니다.");
+          return;
+      }
+      await axios.post(`${process.env.VUE_APP_API_PATH}/api/v1/auth/login`, this.form)
+        .then((res) => {
+          alert("환영합니다!");
+          this.$store.commit("setAccessToken", res.data);
+          this.$store.commit("setUser", this.form.email);
+          this.$router.push('/');
+        })
+        .catch(error => {
+          alert("아이디 또는 비밀번호 확인 후 시도해 주세요.");
+          console.log(error);
+        });
+    },
+    async signup() {
+      if (this.form.email == null || this.form.password == null || this.form.name == null) {
+        alert("입력하지 않은 칸이 있습니다.");
+        return;
+      }
+      else if (this.form.password !== this.password2) {
+        alert("비밀번호가 일치하지 않습니다.");
+        return;
+      }
+      await axios.post(`${process.env.VUE_APP_API_PATH}/api/v1/auth`, this.form)
+        .then(() => {
+          alert("회원가입 성공! 로그인 후 이용해주세요!");
+          this.$router.push('/');
+        })
+        .catch(error => {
+          alert("이미 사용중인 아이디 입니다. 다시 입력해주세요.");
+          console.log(error);
+        });
+    }
+  }
+}
 </script>
 <style lang="scss" scoped>
 section {
-    height: 75vh;
+  height: 80vh;
 }
 
-
 html {
-    height: 100%;
+  height: 100%;
 }
 
 body {
-    margin: 0;
-    padding: 0;
-    font-family: sans-serif;
-    background: linear-gradient(#141e30, #243b55);
+  margin: 0;
+  padding: 0;
+  font-family: sans-serif;
+  background: linear-gradient(#141e30, #243b55);
 }
 
 .login-box {
-    position: absolute;
-    top: 50%;
-    left: 50%;
-    width: 400px;
-    padding: 40px;
-    transform: translate(-50%, -50%);
-    background: rgba(0, 0, 0, .5);
-    box-sizing: border-box;
-    box-shadow: 0 15px 25px rgba(0, 0, 0, .6);
-    border-radius: 10px;
+  position: absolute;
+  top: 50%;
+  left: 50%;
+  width: 400px;
+  padding: 40px;
+  transform: translate(-50%, -50%);
+  background: rgba(0, 0, 0, .2);
+  box-sizing: border-box;
+  box-shadow: 0 15px 25px rgba(0, 0, 0, .3);
+  border-radius: 10px;
 }
 
 .login-box h2 {
-    margin: 0 0 30px;
-    padding: 0;
-    color: #fff;
-    text-align: center;
+  margin: 0 0 30px;
+  padding: 0;
+  color: #fff;
+  text-align: center;
 }
 
 .login-box .user-box {
-    position: relative;
+  position: relative;
 }
 
 .login-box .user-box input {
-    width: 100%;
-    padding: 10px 0;
-    font-size: 16px;
-    color: #fff;
-    margin-bottom: 30px;
-    border: none;
-    border-bottom: 1px solid #fff;
-    outline: none;
-    background: transparent;
+  width: 100%;
+  padding: 10px 0;
+  font-size: 16px;
+  color: #fff;
+  margin-bottom: 30px;
+  border: none;
+  border-bottom: 1px solid #fff;
+  outline: none;
+  background: transparent;
 }
 
 .login-box .user-box label {
-    position: absolute;
-    top: 0;
-    left: 0;
-    padding: 10px 0;
-    font-size: 16px;
-    color: #fff;
-    pointer-events: none;
-    transition: .5s;
+  position: absolute;
+  top: 0;
+  left: 0;
+  padding: 10px 0;
+  font-size: 15px;
+  color: #ccc;
+  pointer-events: none;
+  transition: .5s;
+  text-transform: capitalize;
 }
 
 .login-box .user-box input:focus~label,
 .login-box .user-box input:valid~label {
-    top: -20px;
-    left: 0;
-    color: #03e9f4;
-    font-size: 12px;
+  top: -20px;
+  left: 0;
+  color: #03e9f4;
+  font-size: 12px;
 }
 
-.login-box form a {
-    position: relative;
-    display: inline-block;
-    padding: 10px 20px;
-    color: #00c7d1;
-    font-size: 16px;
-    text-decoration: none;
-    text-transform: uppercase;
-    overflow: hidden;
-    transition: .5s;
-    margin-top: 40px;
-    letter-spacing: 4px;
-    border: 1px solid rgba(255, 255, 255, 0.582);
+.login-btn {
+  background-color: #141e30;
+  position: relative;
+  display: inline-block;
+  padding: 10px 20px;
+  margin: 0 5px;
+  color: #00c7d1;
+  font-size: 15px;
+  text-decoration: none;
+  text-transform: uppercase;
+  overflow: hidden;
+  transition: .5s;
+  margin-top: 20px;
+  letter-spacing: 4px;
+  border: 1px solid rgba(255, 255, 255, 0.582);
 }
 
-.login-box a:hover {
-    color: #fff;
-}
-
-.login-box a span {
-    position: absolute;
-    display: block;
+.login-btn:hover {
+  color: #fff;
 }
 </style>
