+++ client/vue.config.js
@@ -1,3 +1,7 @@
+const path = require('path');
+const dotenv = require('dotenv');
+dotenv.config({ path: path.join(__dirname, '.env') });
+
 const { defineConfig } = require('@vue/cli-service')
 module.exports = {
   // devServer: {
