+++ server/src/main/java/com/crossfit/server/config/jwt/JwtTokenProvider.java
@@ -1,0 +1,155 @@
+package com.crossfit.server.config.jwt;
+
+
+import com.crossfit.server.dto.member.LoginResponseDto;
+import com.crossfit.server.entity.Member;
+import com.crossfit.server.repository.MemberRepository;
+import io.jsonwebtoken.*;
+import io.jsonwebtoken.io.Decoders;
+import io.jsonwebtoken.security.Keys;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.core.userdetails.User;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.stereotype.Component;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.security.Key;
+import java.util.*;
+import java.util.stream.Collectors;
+
+@Slf4j
+@Component
+public class JwtTokenProvider {
+
+    private final MemberRepository memberRepository;
+    private final Key key;
+    private final int ACCESS_TOKEN_SECOND = 60 * 30 * 1000;
+    private long REFRESH_TOKEN_SECOND = 60 * 60 * 24 * 14 * 1000 ;
+
+    public JwtTokenProvider(@Value("${jwt.secret-key}") String secretKey, MemberRepository memberRepository) {
+        this.memberRepository = memberRepository;
+        byte[] keyBytes = Decoders.BASE64.decode(secretKey);
+        this.key = Keys.hmacShaKeyFor(keyBytes);
+    }
+
+    public LoginResponseDto generateToken(Authentication authentication) {
+        String authorities = authentication.getAuthorities().stream()
+                .map(GrantedAuthority::getAuthority)
+                .collect(Collectors.joining(","));
+
+        long now = System.currentTimeMillis();
+
+        Date accessTokenExpiresIn = new Date(now + ACCESS_TOKEN_SECOND);
+        String accessToken = Jwts.builder()
+                .setSubject(authentication.getName())
+                .claim("auth", authorities)
+                .setExpiration(accessTokenExpiresIn)
+                .signWith(key, SignatureAlgorithm.HS256)
+                .compact();
+
+        Date refreshTokenExpiresIn = new Date(now + REFRESH_TOKEN_SECOND);
+        String refreshToken = Jwts.builder()
+                .setExpiration(refreshTokenExpiresIn)
+                .signWith(key, SignatureAlgorithm.HS256)
+                .compact();
+
+        return LoginResponseDto.builder()
+                .grantType("Bearer")
+                .accessToken(accessToken)
+                .refreshToken(refreshToken)
+                .build();
+    }
+
+    @Transactional
+    public String generateAccessToken(String refreshToken) {
+        String accessToken = null;
+
+        Member member = getMemberByRefreshToken(refreshToken);
+        if (member != null) {
+            String username = member.getEmail();
+            List<GrantedAuthority> authorities = member.getAuthorities().stream()
+                    .map(authority -> new SimpleGrantedAuthority(authority.getAuthorityName()))
+                    .collect(Collectors.toList());
+
+            Date accessTokenExpiresIn = new Date(System.currentTimeMillis() + ACCESS_TOKEN_SECOND);
+            String newAccessToken = Jwts.builder()
+                    .setSubject(username)
+                    .claim("auth", authorities.stream()
+                            .map(GrantedAuthority::getAuthority)
+                            .collect(Collectors.toList()))
+                    .setExpiration(accessTokenExpiresIn)
+                    .signWith(key, SignatureAlgorithm.HS256)
+                    .compact();
+
+            accessToken = newAccessToken;
+        }
+
+        return accessToken;
+    }
+
+    public Authentication getAuthentication(String accessToken) {
+        Claims claims = parseClaims(accessToken);
+
+        if (claims.get("auth") == null) {
+            throw new RuntimeException("권한 정보가 없는 토큰입니다.");
+        }
+
+        Collection<? extends GrantedAuthority> authorities =
+                Arrays.stream(claims.get("auth").toString().split(","))
+                        .map(SimpleGrantedAuthority::new)
+                        .collect(Collectors.toList());
+
+        UserDetails principal = new User(claims.getSubject(), "", authorities);
+        return new UsernamePasswordAuthenticationToken(principal, "", authorities);
+    }
+
+    public boolean validateToken(String token) {
+        try {
+            Jwts.parserBuilder().setSigningKey(key).build().parseClaimsJws(token);
+            return true;
+        } catch (io.jsonwebtoken.security.SecurityException | MalformedJwtException e) {
+            log.info("Invalid JWT Token", e);
+        } catch (ExpiredJwtException e) {
+            log.info("Expired JWT Token", e);
+        } catch (UnsupportedJwtException e) {
+            log.info("Unsupported JWT Token", e);
+        } catch (IllegalArgumentException e) {
+            log.info("JWT claims string is empty.", e);
+        }
+        return false;
+    }
+
+    private Claims parseClaims(String accessToken) {
+        try {
+            return Jwts.parserBuilder().setSigningKey(key).build().parseClaimsJws(accessToken).getBody();
+        } catch (ExpiredJwtException e) {
+            return e.getClaims();
+        }
+    }
+
+    public boolean validateRefreshToken(String refreshToken) {
+        try {
+            Jwts.parserBuilder().setSigningKey(key).build().parseClaimsJws(refreshToken);
+            return true;
+        } catch (ExpiredJwtException e) {
+            // refreshToken이 만료된 경우 예외 처리
+            return false;
+        }
+    }
+
+    public Member getMemberByRefreshToken(String refreshToken) {
+        Optional<Member> memberOptional = memberRepository.findByRefreshToken(refreshToken);
+        if (memberOptional.isPresent()) {
+            return memberOptional.get();
+        } else {
+            throw new RuntimeException("멤버를 찾을 수 없습니다.");
+        }
+    }
+
+
+}
