+++ server/src/main/java/com/crossfit/server/controller/IndexController.java
@@ -1,0 +1,15 @@
+package com.crossfit.server.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.ResponseBody;
+
+@Controller
+@ResponseBody
+public class IndexController {
+
+    @GetMapping("/")
+    public String index() {
+        return "Welcome Crossfit Page!";
+    }
+}
