--- server/src/main/java/com/crossfit/server/domain/gym/entity/Board.java
@@ -1,25 +1,0 @@
-package com.crossfit.server.domain.gym.entity;
-
-import com.crossfit.server.domain.member.entity.Member;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import javax.persistence.*;
-
-@Getter
-@Entity
-@NoArgsConstructor
-public class Board {
-
-    @Id
-    @Column(name = "board_id")
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false, length = 30)
-    private String type;
-
-    @ManyToOne
-    @JoinColumn(name = "gym_id")
-    private Gym gym;
-}
