--- server/src/main/java/com/crossfit/server/domain/gym/entity/Gym.java
@@ -1,28 +1,0 @@
-package com.crossfit.server.domain.gym.entity;
-
-import com.crossfit.server.domain.member.entity.Member;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import javax.persistence.*;
-
-@Getter
-@Entity
-@NoArgsConstructor
-public class Gym {
-
-    @Id
-    @Column(name = "gym_id")
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false, unique = true)
-    private String name;
-
-    @Column(nullable = false)
-    private String location;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "member_id")
-    private Member member;
-}
