--- server/src/main/java/com/crossfit/server/domain/gym/entity/Post.java
@@ -1,36 +1,0 @@
-package com.crossfit.server.domain.gym.entity;
-
-import com.crossfit.server.domain.member.entity.Member;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import javax.persistence.*;
-
-
-@Getter
-@Entity
-@NoArgsConstructor
-public class Post {
-
-    @Id
-    @Column(name = "post_id")
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(length = 50)
-    private String title;
-
-    @Column(columnDefinition = "TEXT")
-    private String contnent;
-
-    @Column
-    private boolean isNotice;
-
-    @ManyToOne
-    @JoinColumn(name = "board_id")
-    private Board board;
-
-    @OneToOne
-    @JoinColumn(name = "member_id")
-    private Member member;
-}
