--- server/src/main/java/com/crossfit/server/domain/gym/entity/Reply.java
@@ -1,29 +1,0 @@
-package com.crossfit.server.domain.gym.entity;
-
-import com.crossfit.server.domain.member.entity.Member;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import javax.persistence.*;
-
-@Getter
-@Entity
-@NoArgsConstructor
-public class Reply {
-
-    @Id
-    @Column(name = "reply_id")
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(columnDefinition = "TEXT")
-    private String content;
-
-    @ManyToOne
-    @JoinColumn(name = "post_id")
-    private Post post;
-
-    @OneToOne
-    @JoinColumn(name = "member_id")
-    private Member member;
-}
