--- server/src/main/java/com/crossfit/server/domain/member/dto/LoginRequestDto.java
@@ -1,22 +1,0 @@
-package com.crossfit.server.domain.member.dto;
-
-import com.crossfit.server.domain.member.entity.Member;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@NoArgsConstructor
-public class LoginRequestDto {
-    private String email;
-    private String password;
-
-    public Member toEntity() {
-        return Member.builder()
-                .email(email)
-                .password(password)
-                .build();
-    }
-
-}
