--- server/src/main/java/com/crossfit/server/domain/member/dto/LoginResponseDto.java
@@ -1,16 +1,0 @@
-package com.crossfit.server.domain.member.dto;
-
-import lombok.*;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-@ToString
-public class LoginResponseDto {
-
-    private String grantType;
-    private String accessToken;
-    private String refreshToken;
-}
