+++ server/src/main/java/com/crossfit/server/dto/member/LoginRequestDto.java
@@ -1,0 +1,22 @@
+package com.crossfit.server.dto.member;
+
+import com.crossfit.server.entity.Member;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@NoArgsConstructor
+public class LoginRequestDto {
+    private String email;
+    private String password;
+
+    public Member toEntity() {
+        return Member.builder()
+                .email(email)
+                .password(password)
+                .build();
+    }
+
+}
