+++ server/src/main/java/com/crossfit/server/dto/member/LoginResponseDto.java
@@ -1,0 +1,16 @@
+package com.crossfit.server.dto.member;
+
+import lombok.*;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+@ToString
+public class LoginResponseDto {
+
+    private String grantType;
+    private String accessToken;
+    private String refreshToken;
+}
