+++ server/src/main/java/com/crossfit/server/entity/Authority.java
@@ -1,0 +1,26 @@
+package com.crossfit.server.entity;
+
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.Id;
+import javax.persistence.Table;
+
+@Entity
+@Table(name = "authority")
+@Getter
+@NoArgsConstructor
+public class Authority {
+
+    @Id
+    @Column(name = "authority_name", length = 50)
+    private String authorityName;
+
+    @Builder
+    public Authority(String authorityName) {
+        this.authorityName = authorityName;
+    }
+}
