+++ server/src/main/java/com/crossfit/server/entity/Board.java
@@ -1,0 +1,24 @@
+package com.crossfit.server.entity;
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import javax.persistence.*;
+
+@Getter
+@Entity
+@NoArgsConstructor
+public class Board {
+
+    @Id
+    @Column(name = "board_id")
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(nullable = false, length = 30)
+    private String type;
+
+    @ManyToOne
+    @JoinColumn(name = "gym_id")
+    private Gym gym;
+}
