+++ server/src/main/java/com/crossfit/server/entity/Post.java
@@ -1,0 +1,35 @@
+package com.crossfit.server.entity;
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import javax.persistence.*;
+
+
+@Getter
+@Entity
+@NoArgsConstructor
+public class Post {
+
+    @Id
+    @Column(name = "post_id")
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(length = 50)
+    private String title;
+
+    @Column(columnDefinition = "TEXT")
+    private String content;
+
+    @Column
+    private boolean isNotice;
+
+    @ManyToOne
+    @JoinColumn(name = "board_id")
+    private Board board;
+
+    @OneToOne
+    @JoinColumn(name = "member_id")
+    private Member member;
+}
