+++ server/src/main/java/com/crossfit/server/entity/Reply.java
@@ -1,0 +1,28 @@
+package com.crossfit.server.entity;
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import javax.persistence.*;
+
+@Getter
+@Entity
+@NoArgsConstructor
+public class Reply {
+
+    @Id
+    @Column(name = "reply_id")
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(columnDefinition = "TEXT")
+    private String content;
+
+    @ManyToOne
+    @JoinColumn(name = "post_id")
+    private Post post;
+
+    @OneToOne
+    @JoinColumn(name = "member_id")
+    private Member member;
+}
