--- server/src/main/java/com/crossfit/server/global/config/JwtSecurityConfig.java
@@ -1,23 +1,0 @@
-package com.crossfit.server.global.config;
-
-import com.crossfit.server.global.security.jwt.JwtTokenProvider;
-import com.crossfit.server.global.security.jwt.filter.JwtAuthenticationFilter;
-import org.springframework.security.config.annotation.SecurityConfigurerAdapter;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.web.DefaultSecurityFilterChain;
-import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
-
-public class JwtSecurityConfig extends SecurityConfigurerAdapter<DefaultSecurityFilterChain, HttpSecurity> {
-
-    private JwtTokenProvider tokenProvider;
-
-    public JwtSecurityConfig(JwtTokenProvider tokenProvider) {
-        this.tokenProvider = tokenProvider;
-    }
-
-    @Override
-    public void configure(HttpSecurity http) {
-        JwtAuthenticationFilter customFilter = new JwtAuthenticationFilter(tokenProvider);
-        http.addFilterBefore(customFilter, UsernamePasswordAuthenticationFilter.class);
-    }
-}
