--- server/src/main/java/com/crossfit/server/global/security/jwt/controller/JwtController.java
@@ -1,36 +1,0 @@
-package com.crossfit.server.global.security.jwt.controller;
-
-import com.crossfit.server.domain.member.dto.LoginResponseDto;
-import com.crossfit.server.global.security.jwt.service.JwtService;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import javax.servlet.http.Cookie;
-import javax.servlet.http.HttpServletRequest;
-
-@RestController
-@RequiredArgsConstructor
-@RequestMapping("/api/v1/jwt")
-public class JwtController {
-
-    private final JwtService jwtService;
-
-    @PostMapping(value = "")
-    public ResponseEntity<LoginResponseDto> refreshToken(HttpServletRequest request) {
-        Cookie[] cookies = request.getCookies();
-        String refreshToken = null;
-
-        if (cookies != null) {
-            for (Cookie cookie : cookies) {
-                if (cookie.getName().equals("refreshToken")) {
-                    refreshToken = cookie.getValue();
-                    break;
-                }
-            }
-        }
-        return jwtService.reToken(refreshToken);
-    }
-}
