+++ server/src/main/resources/application.properties
@@ -1,9 +1,14 @@
 server.port = 9959
 
-spring.datasource.driverClassName=org.mariadb.jdbc.Driver
+spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
 spring.datasource.url = ${DB_URL}
 spring.datasource.username = ${DB_USERNAME}
 spring.datasource.password = ${DB_PASSWORD}
+spring.sql.init.encoding=UTF-8
+spring.sql.init.mode=always
+
+spring.jpa.database=mysql
+spring.jpa.show-sql=true
 spring.jpa.hibernate.ddl-auto=update
 spring.jpa.defer-datasource-initialization=true
 spring.mvc.hiddenmethod.filter.enabled=true
